Building A ML Model for classifying song
Classifying a English Song in Happy or Sad Category .

What is sentiment analysis?
Sentiment Analysis is: The process of determining the emotional tone behind a series of words, used to gain an understanding of the the attitudes, opinions and emotions expressed within an online mention.

Why do we use sentiment analysis ?
Sentiment analysis is extremely useful in social media monitoring as it allows us to gain an overview of the wider public opinion behind certain topics.The applications of sentiment analysis are broad and powerful. The ability to extract insights from social data is a practice that is being widely adopted by organisations across the world.